Add files and directory framework to generate the xCAT-confluent RPM
Cleaned up some of the comments in the beginning of buildcore.sh
using spaces instead of tabs and added the build for xCAT-confluent.

The xCAT-confluent rpm is intended to be used for xCAT2.X releases to
support confluent.  This rpm has a requirement on confluent_server package.
For 2.9.1, xCAT-confluent is optional.  For 2.10, the plan is to default
the console service to confluent.