Change wvid to pass IP to client
When remote xCAT client is used, often the client will not have
the same name resolution in place as the management node.  For
wvid, mitigate this by passing the IP address to the client.
If name resolution worked before, this should be no downside.
If routing also doesn't work, well the user is no worse off than
before at the moment.