Improve IPMI FRU parsing for Lenovo Thinkservers
Lenovo Thinkservers provide logically addressed FRUs that
aren't typed as logical FRU.  Change so that the
addressing rather than the type drives our behavior to
read or skip.

Thinkservers explicitly have a locator record for FRU 0.
Ignore such a locator record.

If a board info area is not followed by a product or
multirecord area, give the board area parsing code access
to the end of the area in case board info data runs
beyond the end of the claimed area.