Restore backwards compatibility to !remote!
A change was made to allow '!remote!' to force indicate a NIC to be
relevant in case it wouldn't naturally occur.  This broke existing
configurations where the NIC would naturally be 'active'.  Attempt
to address by restoring the old match and enabling the new code
when an interface name is detected after '!remote!', but otherwise
act like pre-2.9.1 dhcp.pm