Rework traffic control to not involve UDP every SSL connection
If the UDP monitor was busy, the SSL connections could back up.
Rework it so that the UDP monitor would get hung up by the SSL
process instead, if *any* of them get backed up.  However, the
nature of the loop in the SSL client is sufficiently lightweight
that it shouldn't hang up the UDP process under any reasonable
circumstance.  If it did hang it up though, it'll be less hung up
than discovery does already.