Merge the pages for Large Cluster and Hierarchical
Realized that the large cluster support and the advanced hierarchical pages are the same.
This initial commit does some refactoring of the pages.

Left around the define_service_node.rst page but it's not in the index.
We need to verify that the content is correct in define_service_nodes.rst