Improve port change check
When changing the port when indicating NCSI
package, the comparison was inadequate.  Improve
things by at least checking the shared/dedicated portion.

Note that it is not comparing the package id, which means
things are imperfect, though they are better than before.
However, it's not too shabby because the NIC being active
is the important part, the NCSI package change happening
later is ok, since the net config is still being put on
the correct IMM Nic.