Add support for a default MTU on a network
Add a MTU column to the networks table.  If set, this will configure the MTU in
DHCP so that nodes will have the correct MTU during provisioning.  This is
important for large systems to enable jumbo frames to reduce CPU usage when
copying diskless images over the network (before the confignics postscript can
run to set the desired MTU).