Ensure the requested kernel version and associated packages are installed
If a specific version of the kernel was requested, make sure any other
kernel-related packages requested in a package list get the same version.
Note, packages that get pulled in due to dependencies are NOT checked.
Also avoid updating the kernel packages when the requested version has already
been installed.