Add extra nics to link-up detection search list
When a bonded NIC configuration is to be used as the primary interface
for a node, often that interface cannot be used to PXE boot the node.
In this case, an different NIC is used as the "installnic".
It's possible that after the node reboots, that the "installnic"
is not configured with an address. The IP address can be moved to the
bonded connection after provisioning.
The additional interfaces used by the node are listed in "NICIPS"
from the "nics" table. This code change parses "NICIPS" for adapter
names and adds them to the INSTALLNIC variable so that all defined
interfaces for this node are checked in the link-up check.
This should still fix "defect 4616"
https://sourceforge.net/p/xcat/bugs/4616/
The original code change was intended to delay the execution of the
xcatinstallpost script until a real network interface came up (i.e. not usb0)