Redirect progress message to rflash log files when updating firmware
Previously rflash process will be terminated when the tty session is closed
which is a very dangerous operation that may bricked the BMC. This patch will
ignore the `int`, `term` and `hup` signal when rflash is running and redirect
the output of ipmitool command to `/var/log/xcat/rflash/<node.log>` which make
the progress trackable. In addition,  -V option is provided to print more
verbose message from ipmitool when `hpm upgrade` command is running.

This patch also update the node status in the `nodelist` table to record
the rflash status.

fix-issue: #1849