Since the genesis kernel is not intended to be long running, there's no need to start ntp daemon, simply attempt to sync with ntpdate.
There's an issue in this code where we query ntpq for the offset too quickly after starting the ntp daemon.  This causes the offset to always be 0 and we continue.  We have seen situations where ntpd starts before the first check and becomes > 1000.  Then it takes 15 minutes before the 2nd attempt and delays the hardware discovery process