Fix timeout issue if bmc is unreachable for rflash
As the original ipmi code do not exit with error code, it is
not possible to judge the success or not from return code of child
procss. This patch will send message from child processes directly.
This patch also change the wait time to check whether the ipmi
session is useable and add power on at the end of rflash process.

close-issue: #2693
close-issue: #2682