added support for copying multiple images simultaneously
packimage.pm now associates a unique port with an image, which is passed
to xcatroot, so uftpd knows on which port to listen.  The port is added
to the request to uftp-listener, which parses it and starts the transfer
only to nodes listening on that port.  Tested on crest1 and crest2 here.