List console status with makegocons -q
Add `-q|--query [noderange]` for makegocons to list the console
status and the host server. If node is not registered in
goconserver, the status will be `unregistered`.

If console releated attribute is not defined for the specified
node but it is included in the noderange, error message for that
node will be printed at first.