Allow restricting service node pools
Service node pools would go as far as the network allowed.
Honor the servicenode attribute to limit visibility to
the service nodes that are relevant and any owner of a dynamic
range.

Also allow execution if the dhcp server is not running (e.g. shutting
down dhcp server on the master node to allow only the service nodes
to provide services).