Fix syncfiles merge update records
The change to fix superset problem also reversed logic
to prefer the 'old' entries to the 'new' entries.  For now,
restore the logic of merge entries superseding existing entries.

What is really warranted is a facility to intelligently sync the account
databases.  Specifically, read in /etc/login.defs, look at UID_MIN and
GID_MIN for the platform.  For UID/GID lower than that, rpms may manage
and such care needs to be given to not change UID/GID, but still need to
update things like shadow and group *membership*, but presrve uid/gid.

Above UID_MIN/GID_MIN, they are user groups and the uid/gid portion is
important in shared filesystem context, so it is imperative that they
work.