Mitigate risk of kdump enable process problems
During kdump enablement, it was potentially mounting under /tmp
or /var/tmp.  This is a very risky move as at the same time, it
is common to have boot activity 'cleaning' tmp.  Move the
mount point elsewhere.

Additionally, try making it RO, and deferring the two write options
to be done by user.