/sbin/mksquashfs does not understand -le and -be
... at least on CentOS 7.7, it doesn't. Building squashfs images thus fails with "mksquashfs could not be run successfully", missing the real reason why not.
As I understand it, mksquashfs should build little or big endian .sfs files according to the architecture it runs on. If the -le / -be switches need to remain, eg. for the sake of cross-architecture or POWER builds: perhaps some experienced xcat programmer can look over the check for "$osver" to be extended to cover CentOS as well (emptying $flags as it does for rhels already).