Update bmcsetup
Don't check $LOCKEDUSERS at all, instead fall back to userslot 2 as a last resort if the user doesn't already exist, and the automatic slot detection doesn't find an unlocked slot.  This is essentially what was happening in the older version of the script before #6380 as $LOCKEDUSERS was only ever referenced here.

Also move setting the username before setting privileges and access levels, as those steps fail if done before username is set for the slot.