Correctly specify initrd file for all netboot configs.
Previously the code which created the netboot config files was
incorrectly specifying the initrd file name in a number of places,
particularly in the xcat/xnba/nets/* files, using a hard-coded `.gz`
suffix even when a `.lzma` file had been created. This change ensures
that the initrd file name determined earlier in the code (either when
the image is created, or in the configonly case by checking the
filesystem) is used correctly when creating the netboot config files.